requirements.txt
kiteconnect>=4.0.0
requests>=2.28.0
pyotp>=2.8.0
python-dotenv>=1.0.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1
websocket-client>=1.7.0

# Backtesting & Data Processing
polars>=0.20.0
numpy>=1.24.0

# Risk Management & Performance Metrics
empyrical-reloaded>=0.5.9
quantstats>=0.0.62
scipy>=1.10.0

# Portfolio Optimization
riskfolio-lib>=5.1.0
PyPortfolioOpt>=1.5.5

# Progress Bars & UI
rich>=13.0.0
tqdm>=4.65.0

# Additional utilities
numba>=0.58.0